requirements.txt
# ctransformers==0.2.5
# sentence-transformers==2.2.2
pinecone-client
langchain_huggingface
langchain
langchain-community
flask
pypdf
python-dotenv
ipykernel
langchain_pinecone
-e .